Apply coding norms

+++ l95/test/lorenz95/GomL95.cc
@@ -13,13 +13,13 @@
 
 #include "./TestConfig.h"
 #include "eckit/config/LocalConfiguration.h"
-#include "oops/base/Variables.h"
 #include "lorenz95/GomL95.h"
 #include "lorenz95/LocsL95.h"
 #include "lorenz95/ObservationL95.h"
 #include "lorenz95/ObsTable.h"
-#include "util/DateTime.h"
+#include "oops/base/Variables.h"
 #include "test/TestFixture.h"
+#include "util/DateTime.h"
 
 namespace test {
 
@@ -55,7 +55,7 @@ BOOST_FIXTURE_TEST_SUITE(test_GomL95, GomTestFixture)
 // -----------------------------------------------------------------------------
   BOOST_AUTO_TEST_CASE(test_GomL95_nobs) {
     boost::scoped_ptr<lorenz95::GomL95> gom(new lorenz95::GomL95(*locs_, *novar_));
-    int ref = 160;
+    size_t ref = 160;
     BOOST_CHECK_EQUAL(gom->size(), ref);
   }
 // -----------------------------------------------------------------------------
